[#feature] 구글 사용자의 정보 DTO 생성

+++ src/main/java/com/petpick/model/ClientTokenResponse.java
@@ -2,12 +2,14 @@ package com.petpick.model;
 
 import lombok.Data;
 
+import java.time.LocalDateTime;
+
 @Data
 public class ClientTokenResponse {
 
     private String accessToken;
 
-    private int expiresIn;
+    private Long expiresIn;
 
     private String tokenType;
 }
